+++ allinpy/__init__.py
@@ -1,5 +1,8 @@
 from . import plotting, utils
 from .io_utils.file_utils import (get_df_subj, get_file_paths, load_data,
                                   sorted_nicely)
-from .plotting.plot_utils import cm2inch, label_subplots, latex_plt
+from .plotting.plot_utils import (center_x, center_y, cm2inch, get_text_coords,
+                                  label_subplots, latex_plt, plot_arrow,
+                                  plot_centered_text, plot_image, plot_rec,
+                                  plot_table)
 from .utils.utilities import callback, safe_div